Projet_2/requirements.txt
csv
pandas
matplotlib.pyplot
sqlite3
statsmodels.api
scipy.stats
seaborn
sklearn.model_selection
sklearn.linear_model
sklearn.cluster
matplotlib
numpy
scikit-learn
scipy
statsmodels
